refactor: modularize server components
