+++ server/routers/__init__.py
@@ -1,0 +1,2 @@
+from . import foods, meals, presets
+__all__ = ["foods", "meals", "presets"]
